refactor: move i18n to separate module
