+++ src/modules/i18n/const/index.ts
