+++ src/modules/shared/const/index.ts
@@ -1,4 +1,3 @@
 export * from './form';
-export * from './i18n';
 export * from './routes';
 export * from './sanitize';
