+++ tsconfig.json
@@ -5,6 +5,7 @@
       "@auth/*": ["modules/auth/*"],
       "@config/*": ["modules/config/*"],
       "@home/*": ["modules/home/*"],
+      "@i18n/*": ["modules/i18n/*"],
       "@navigation/*": ["modules/navigation/*"],
       "@reading-room/*": ["modules/reading-room/*"],
       "@shared/*": ["modules/shared/*"]
